+++ src/main/java/com/concord/petmily/vaccination/service/VaccinationService.java
@@ -1,46 +1,24 @@
 package com.concord.petmily.vaccination.service;
 
-import com.concord.petmily.vaccination.repository.VaccinationRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+public interface VaccinationService {
 
-/**
- *  접종 비즈니스 로직
- *
- */
-@Service
-@RequiredArgsConstructor
-public class VaccinationService {
+    /**
+     * 접종 등록
+     */
+    void createPet();
 
-  private final VaccinationRepository vaccinationRepository;
+    /**
+     * 접종 조회
+     */
+    void userPetList();
 
-  /**
-   * 접종 등록
-   */
-  public void createPet(){
+    /**
+     * 접종 정보 수정
+     */
+    void modifierPet();
 
-  }
-
-  /**
-   * 접종 조회
-   */
-  public void userPetList(){
-
-  }
-
-  /**
-   * 접종 정보 수정
-   */
-  public void modifierPet(){
-
-  }
-
-  /**
-   * 접종 정보 삭제 처리
-   */
-
-  public void deletePet(){
-
-  }
-
+    /**
+     * 접종 정보 삭제 처리
+     */
+    void deletePet();
 }
